+++ src/platform/posix/posix_tcpconn.c
@@ -66,7 +66,7 @@ tcp_dowrite(nni_tcp_conn *c)
 		hdr.msg_iovlen = niov;
 		hdr.msg_iov    = iovec;
 
-		if ((n = sendmsg(fd, &hdr, MSG_NOSIGNAL)) < 0) {
+		if ((n = (int) sendmsg(fd, &hdr, MSG_NOSIGNAL)) < 0) {
 			switch (errno) {
 			case EINTR:
 				continue;
@@ -128,7 +128,7 @@ tcp_doread(nni_tcp_conn *c)
 			}
 		}
 
-		if ((n = readv(fd, iovec, niov)) < 0) {
+		if ((n = (int) readv(fd, iovec, niov)) < 0) {
 			switch (errno) {
 			case EINTR:
 				continue;
